Updated for python 3.5
I don't know if this is exactly wanted, but i wanted to get this working with python 3.5
Few small changes I had to make. One was where urllib.quote came from. In python 3.5 it comes from a different parse in urllib
second was i had issues trying to base64encoded the string. By following the API instructions, I have used the alternative approach of putting the client id and client secret into the code payload and removing the headers.
"An alternative way to send the client id and secret is as request parameters (client_id and client_secret) in the POST body, instead of sending them base64-encoded in the header."
https://beta.developer.spotify.com/documentation/general/guides/authorization-guide/